+++ packages/cinterop/src/jvm/kotlin/io/realm/internal/interop/RealmInterop.kt
@@ -26,13 +26,12 @@ import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.CoroutineStart
 import kotlinx.coroutines.launch
-import java.lang.reflect.Method
 
 // FIXME API-CLEANUP Rename io.realm.interop. to something with platform?
 //  https://github.com/realm/realm-kotlin/issues/56
 
-private val INVALID_CLASS_KEY: Long by lazy { realmc.getRLM_INVALID_CLASS_KEY() }
-private val INVALID_PROPERTY_KEY: Long by lazy { realmc.getRLM_INVALID_PROPERTY_KEY() }
+actual val INVALID_CLASS_KEY: ClassKey by lazy { ClassKey(realmc.getRLM_INVALID_CLASS_KEY()) }
+actual val INVALID_PROPERTY_KEY: PropertyKey by lazy { PropertyKey(realmc.getRLM_INVALID_PROPERTY_KEY()) }
 
 /**
  * JVM/Android interop implementation.
@@ -42,21 +41,6 @@ private val INVALID_PROPERTY_KEY: Long by lazy { realmc.getRLM_INVALID_PROPERTY_
  */
 @Suppress("LargeClass", "FunctionNaming", "TooGenericExceptionCaught")
 actual object RealmInterop {
-
-    // TODO API-CLEANUP Maybe pull library loading into separate method
-    //  https://github.com/realm/realm-kotlin/issues/56
-    init {
-        if (isAndroid()) {
-            System.loadLibrary("realmc")
-        } else {
-            // otherwise locate, using reflection, the dependency SoLoader and call load
-            // (calling SoLoader directly will create a circular dependency with `jvmMain`)
-            val classToLoad = Class.forName("io.realm.jvm.SoLoader")
-            val instance = classToLoad.newInstance()
-            val loadMethod: Method = classToLoad.getDeclaredMethod("load")
-            loadMethod.invoke(instance)
-        }
-    }
 
     actual fun realm_get_version_id(realm: NativePointer): Long {
         val version = realm_version_id_t()
@@ -79,21 +63,21 @@ actual object RealmInterop {
         return result.first()
     }
 
-    actual fun realm_schema_new(tables: List<Table>): NativePointer {
-        val count = tables.size
+    actual fun realm_schema_new(schema: List<Pair<ClassInfo, List<PropertyInfo>>>): NativePointer {
+        val count = schema.size
         val cclasses = realmc.new_classArray(count)
         val cproperties = realmc.new_propertyArrayArray(count)
 
-        for ((i, clazz) in tables.withIndex()) {
-            val properties = clazz.properties
+        for ((i, entry) in schema.withIndex()) {
+            val (clazz, properties) = entry
             // Class
             val cclass = realm_class_info_t().apply {
                 name = clazz.name
                 primary_key = clazz.primaryKey ?: ""
                 num_properties = properties.size.toLong()
                 num_computed_properties = 0
-                key = INVALID_CLASS_KEY
-                flags = clazz.flags.fold(0) { flags, element -> flags or element.nativeValue }
+                key = INVALID_CLASS_KEY.key
+                flags = clazz.flags
             }
             // Properties
             val classProperties = realmc.new_propertyArray(properties.size)
@@ -105,8 +89,8 @@ actual object RealmInterop {
                     collection_type = property.collectionType.nativeValue
                     link_target = property.linkTarget
                     link_origin_property_name = property.linkOriginPropertyName
-                    key = INVALID_PROPERTY_KEY
-                    flags = property.flags.fold(0) { flags, element -> flags or element.nativeValue }
+                    key = INVALID_PROPERTY_KEY.key
+                    flags = property.flags
                 }
                 realmc.propertyArray_setitem(classProperties, j, cproperty)
             }
@@ -192,6 +176,59 @@ actual object RealmInterop {
         return realmc.realm_get_num_classes((realm as LongPointerWrapper).ptr)
     }
 
+    actual fun realm_get_class_keys(realm: NativePointer): List<ClassKey> {
+        val count = realm_get_num_classes(realm)
+        val keys = LongArray(count.toInt())
+        val outCount = longArrayOf(0)
+        realmc.realm_get_class_keys(realm.cptr(), keys, count, outCount)
+        if (count != outCount[0]) {
+            error("Invalid schema: Insufficient keys; got ${outCount[0]}, expected $count")
+        }
+        return keys.map { ClassKey(it) }
+    }
+
+    actual fun realm_find_class(realm: NativePointer, name: String): ClassKey? {
+        val info = realm_class_info_t()
+        val found = booleanArrayOf(false)
+        realmc.realm_find_class((realm as LongPointerWrapper).ptr, name, found, info)
+        return if (found[0]) {
+            ClassKey(info.key)
+        } else {
+            null
+        }
+    }
+
+    actual fun realm_get_class(realm: NativePointer, classKey: ClassKey): ClassInfo {
+        val info = realm_class_info_t()
+        realmc.realm_get_class(realm.cptr(), classKey.key, info)
+        return with(info) {
+            ClassInfo(name, primary_key, num_properties, num_computed_properties, ClassKey(key), flags)
+        }
+    }
+    actual fun realm_get_class_properties(realm: NativePointer, classKey: ClassKey, max: Long): List<PropertyInfo> {
+        val properties = realmc.new_propertyArray(max.toInt())
+        val outCount = longArrayOf(0)
+        realmc.realm_get_class_properties(realm.cptr(), classKey.key, properties, max, outCount)
+        return if (outCount[0] > 0) {
+            (0 until outCount[0]).map { i ->
+                with(realmc.propertyArray_getitem(properties, i.toInt())) {
+                    PropertyInfo(
+                        name,
+                        public_name,
+                        PropertyType.from(type),
+                        CollectionType.from(collection_type),
+                        link_target,
+                        link_origin_property_name,
+                        PropertyKey(key),
+                        flags
+                    )
+                }
+            }
+        } else {
+            emptyList()
+        }
+    }
+
     actual fun realm_release(p: NativePointer) {
         realmc.realm_release((p as LongPointerWrapper).ptr)
     }
@@ -242,26 +279,20 @@ actual object RealmInterop {
         }
     }
 
-    actual fun realm_find_class(realm: NativePointer, name: String): ClassKey {
-        val info = realm_class_info_t()
-        val found = booleanArrayOf(false)
-        realmc.realm_find_class((realm as LongPointerWrapper).ptr, name, found, info)
-        if (!found[0]) {
-            throw IllegalArgumentException("Cannot find class: '$name")
-        }
-        return ClassKey(info.key)
-    }
-
     actual fun realm_object_as_link(obj: NativePointer): Link {
         val link: realm_link_t = realmc.realm_object_as_link(obj.cptr())
-        return Link(link.target_table, link.target)
+        return Link(ClassKey(link.target_table), link.target)
     }
 
-    actual fun realm_get_col_key(realm: NativePointer, table: String, col: String): ColumnKey {
-        return ColumnKey(propertyInfo(realm, classInfo(realm, table), col).key)
+    actual fun realm_get_col_key(
+        realm: NativePointer,
+        className: String,
+        col: String
+    ): PropertyKey {
+        return PropertyKey(propertyInfo(realm, classInfo(realm, className), col).key)
     }
 
-    actual fun <T> realm_get_value(obj: NativePointer, key: ColumnKey): T {
+    actual fun <T> realm_get_value(obj: NativePointer, key: PropertyKey): T {
         // TODO OPTIMIZED Consider optimizing this to construct T in JNI call
         val cvalue = realm_value_t()
         realmc.realm_get_value((obj as LongPointerWrapper).ptr, key.key, cvalue)
@@ -292,13 +323,18 @@ actual object RealmInterop {
         } as T
     }
 
-    actual fun <T> realm_set_value(o: NativePointer, key: ColumnKey, value: T, isDefault: Boolean) {
+    actual fun <T> realm_set_value(o: NativePointer, key: PropertyKey, value: T, isDefault: Boolean) {
         val cvalue = to_realm_value(value)
         realmc.realm_set_value((o as LongPointerWrapper).ptr, key.key, cvalue, isDefault)
     }
 
-    actual fun realm_get_list(obj: NativePointer, key: ColumnKey): NativePointer {
-        return LongPointerWrapper(realmc.realm_get_list((obj as LongPointerWrapper).ptr, key.key))
+    actual fun realm_get_list(obj: NativePointer, key: PropertyKey): NativePointer {
+        return LongPointerWrapper(
+            realmc.realm_get_list(
+                (obj as LongPointerWrapper).ptr,
+                key.key
+            )
+        )
     }
 
     actual fun realm_list_size(list: NativePointer): Long {
@@ -610,12 +646,12 @@ actual object RealmInterop {
         realmc.realm_config_set_sync_config(realmConfiguration.cptr(), syncConfiguration.cptr())
     }
 
-    private fun classInfo(realm: NativePointer, table: String): realm_class_info_t {
+    private fun classInfo(realm: NativePointer, className: String): realm_class_info_t {
         val found = booleanArrayOf(false)
         val classInfo = realm_class_info_t()
-        realmc.realm_find_class((realm as LongPointerWrapper).ptr, table, found, classInfo)
+        realmc.realm_find_class((realm as LongPointerWrapper).ptr, className, found, classInfo)
         if (!found[0]) {
-            throw IllegalArgumentException("Cannot find class: '$table")
+            throw IllegalArgumentException("Cannot find class: '$className")
         }
         return classInfo
     }
@@ -630,9 +666,9 @@ actual object RealmInterop {
         return pinfo
     }
 
-    actual fun realm_query_parse(realm: NativePointer, table: String, query: String, vararg args: Any?): NativePointer {
+    actual fun realm_query_parse(realm: NativePointer, className: String, query: String, vararg args: Any?): NativePointer {
         val count = args.size
-        val classKey = classInfo(realm, table).key
+        val classKey = classInfo(realm, className).key
         val cArgs = realmc.new_valueArray(count)
         args.mapIndexed { i, arg ->
             realmc.valueArray_setitem(cArgs, i, to_realm_value(arg))
@@ -672,7 +708,7 @@ actual object RealmInterop {
     }
 
     actual fun realm_get_object(realm: NativePointer, link: Link): NativePointer {
-        return LongPointerWrapper(realmc.realm_get_object(realm.cptr(), link.tableKey, link.objKey))
+        return LongPointerWrapper(realmc.realm_get_object(realm.cptr(), link.classKey.key, link.objKey))
     }
 
     actual fun realm_object_find_with_primary_key(realm: NativePointer, classKey: ClassKey, primaryKey: Any?): NativePointer? {
@@ -712,7 +748,7 @@ actual object RealmInterop {
         if (this.type != realm_value_type_e.RLM_TYPE_LINK) {
             error("Value is not of type link: $this.type")
         }
-        return Link(this.link.target_table, this.link.target)
+        return Link(ClassKey(this.link.target_table), this.link.target)
     }
 }