+++ packages/library-base/src/commonMain/kotlin/io/realm/BaseRealm.kt
@@ -14,6 +14,8 @@
  */
 package io.realm
 
+import io.realm.schema.RealmSchema
+
 /**
  * Base class for all Realm instances ([Realm] and [MutableRealm]).
  */
@@ -22,6 +24,13 @@ interface BaseRealm : Versioned {
      * Configuration used to configure this Realm instance.
      */
     val configuration: Configuration
+
+    /**
+     * Returns an immutable schema of the realm.
+     *
+     * @return the schema of the realm.
+     */
+    fun schema(): RealmSchema
 
     /**
      * Returns the current number of active versions in the Realm file. A large number of active versions can have