+++ packages/library-base/src/commonMain/kotlin/io/realm/schema/RealmStorageType.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Realm Inc.
+ * Copyright 2021 Realm Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,14 +14,17 @@
  * limitations under the License.
  */
 
-package io.realm.internal.interop
+package io.realm.schema
 
-// FIXME API-SCHEMA Platform independent class definition. Maybe rework into utility method called in Realm
-//  object's companion schema mechanism depending on how we relate this to the actual schema/runtime
-//  realm_class_info_t.
-data class Table(
-    val name: String,
-    val primaryKey: String?,
-    val flags: Set<ClassFlag> = setOf(ClassFlag.RLM_CLASS_NORMAL),
-    val properties: List<Property>
-)
+/**
+ * The various types that are used when storing the property values in the realm.
+ */
+enum class RealmStorageType {
+    BOOL,
+    INT,
+    STRING,
+    OBJECT,
+    FLOAT,
+    DOUBLE,
+    TIMESTAMP;
+}