+++ packages/plugin-compiler/src/main/kotlin/io/realm/compiler/Identifiers.kt
@@ -32,7 +32,7 @@ internal object Names {
     val OBJECT_POINTER = Name.identifier("${REALM_SYNTHETIC_PROPERTY_PREFIX}ObjectPointer")
     // names must match `RealmObjectInternal` properties
     val REALM_OWNER = Name.identifier("${REALM_SYNTHETIC_PROPERTY_PREFIX}Owner")
-    val OBJECT_TABLE_NAME = Name.identifier("${REALM_SYNTHETIC_PROPERTY_PREFIX}TableName")
+    val OBJECT_CLASS_NAME = Name.identifier("${REALM_SYNTHETIC_PROPERTY_PREFIX}TableName")
     val OBJECT_IS_MANAGED = Name.identifier("${REALM_SYNTHETIC_PROPERTY_PREFIX}IsManaged")
     val MEDIATOR = Name.identifier("${REALM_SYNTHETIC_PROPERTY_PREFIX}Mediator")
 
@@ -47,11 +47,8 @@ internal object Names {
     val REALM_OBJECT_HELPER_SET_LIST = Name.identifier("setList")
 
     // Schema related names
-    val CLASS_FLAG_NORMAL = Name.identifier("RLM_CLASS_NORMAL")
-    val PROPERTY_FLAG_NORMAL = Name.identifier("RLM_PROPERTY_NORMAL")
-    val PROPERTY_FLAG_NULLABLE = Name.identifier("RLM_PROPERTY_NULLABLE")
-    val PROPERTY_FLAG_PRIMARY_KEY = Name.identifier("RLM_PROPERTY_PRIMARY_KEY")
-    val PROPERTY_FLAG_INDEX = Name.identifier("RLM_PROPERTY_INDEXED")
+    val CLASS_INFO_CREATE = Name.identifier("create")
+    val PROPERTY_INFO_CREATE = Name.identifier("create")
     val PROPERTY_TYPE_OBJECT = Name.identifier("RLM_PROPERTY_TYPE_OBJECT")
     val PROPERTY_COLLECTION_TYPE_NONE = Name.identifier("RLM_COLLECTION_TYPE_NONE")
     val PROPERTY_COLLECTION_TYPE_LIST = Name.identifier("RLM_COLLECTION_TYPE_LIST")
@@ -74,6 +71,7 @@ internal object FqNames {
     val REALM_OBJECT_HELPER = FqName("io.realm.internal.RealmObjectHelper")
     val REALM_REFERENCE = FqName("io.realm.internal.RealmReference")
     val REALM_MEDIATOR_INTERFACE = FqName("io.realm.internal.Mediator")
+    val REALM_CLASS_IMPL = FqName("io.realm.internal.schema.RealmClassImpl")
     val REALM_CONFIGURATION = FqName("io.realm.RealmConfiguration")
     val SYNC_CONFIGURATION = FqName("io.realm.mongodb.SyncConfiguration")
     val REALM_CONFIGURATION_BUILDER = FqName("io.realm.RealmConfiguration.Builder")
@@ -88,12 +86,10 @@ internal object FqNames {
     val KOTLIN_COLLECTIONS_MAPOF = FqName("kotlin.collections.mapOf")
     val KOTLIN_REFLECT_KPROPERTY1 = FqName("kotlin.reflect.KMutableProperty1")
     // Schema related types
-    val TABLE = FqName("io.realm.internal.interop.Table")
-    val CLASS_FLAG = FqName("io.realm.internal.interop.ClassFlag")
-    val PROPERTY = FqName("io.realm.internal.interop.Property")
+    val CLASS_INFO = FqName("io.realm.internal.interop.ClassInfo")
+    val PROPERTY_INFO = FqName("io.realm.internal.interop.PropertyInfo")
     val PROPERTY_TYPE = FqName("io.realm.internal.interop.PropertyType")
     val COLLECTION_TYPE = FqName("io.realm.internal.interop.CollectionType")
-    val PROPERTY_FLAG = FqName("io.realm.internal.interop.PropertyFlag")
     val PRIMARY_KEY_ANNOTATION = FqName("io.realm.annotations.PrimaryKey")
     val INDEX_ANNOTATION = FqName("io.realm.annotations.Index")
     val IGNORE_ANNOTATION = FqName("io.realm.annotations.Ignore")