+++ packages/plugin-compiler/src/main/kotlin/io/realm/compiler/RealmModelSyntheticPropertiesGeneration.kt
@@ -16,12 +16,11 @@
 
 package io.realm.compiler
 
-import io.realm.compiler.FqNames.CLASS_FLAG
+import io.realm.compiler.FqNames.CLASS_INFO
 import io.realm.compiler.FqNames.COLLECTION_TYPE
 import io.realm.compiler.FqNames.INDEX_ANNOTATION
 import io.realm.compiler.FqNames.PRIMARY_KEY_ANNOTATION
-import io.realm.compiler.FqNames.PROPERTY
-import io.realm.compiler.FqNames.PROPERTY_FLAG
+import io.realm.compiler.FqNames.PROPERTY_INFO
 import io.realm.compiler.FqNames.PROPERTY_TYPE
 import io.realm.compiler.FqNames.REALM_INSTANT
 import io.realm.compiler.FqNames.REALM_MEDIATOR_INTERFACE
@@ -29,17 +28,14 @@ import io.realm.compiler.FqNames.REALM_MODEL_COMPANION
 import io.realm.compiler.FqNames.REALM_NATIVE_POINTER
 import io.realm.compiler.FqNames.REALM_OBJECT_INTERNAL_INTERFACE
 import io.realm.compiler.FqNames.REALM_REFERENCE
-import io.realm.compiler.FqNames.TABLE
-import io.realm.compiler.Names.CLASS_FLAG_NORMAL
+import io.realm.compiler.Names.CLASS_INFO_CREATE
 import io.realm.compiler.Names.MEDIATOR
+import io.realm.compiler.Names.OBJECT_CLASS_NAME
 import io.realm.compiler.Names.OBJECT_IS_MANAGED
 import io.realm.compiler.Names.OBJECT_POINTER
-import io.realm.compiler.Names.OBJECT_TABLE_NAME
 import io.realm.compiler.Names.PROPERTY_COLLECTION_TYPE_LIST
 import io.realm.compiler.Names.PROPERTY_COLLECTION_TYPE_NONE
-import io.realm.compiler.Names.PROPERTY_FLAG_INDEX
-import io.realm.compiler.Names.PROPERTY_FLAG_NULLABLE
-import io.realm.compiler.Names.PROPERTY_FLAG_PRIMARY_KEY
+import io.realm.compiler.Names.PROPERTY_INFO_CREATE
 import io.realm.compiler.Names.PROPERTY_TYPE_OBJECT
 import io.realm.compiler.Names.REALM_OBJECT_COMPANION_FIELDS_MEMBER
 import io.realm.compiler.Names.REALM_OBJECT_COMPANION_NEW_INSTANCE_METHOD
@@ -60,8 +56,11 @@ import org.jetbrains.kotlin.ir.builders.declarations.addProperty
 import org.jetbrains.kotlin.ir.builders.declarations.addValueParameter
 import org.jetbrains.kotlin.ir.builders.declarations.buildField
 import org.jetbrains.kotlin.ir.builders.irBlockBody
+import org.jetbrains.kotlin.ir.builders.irBoolean
 import org.jetbrains.kotlin.ir.builders.irGet
 import org.jetbrains.kotlin.ir.builders.irGetField
+import org.jetbrains.kotlin.ir.builders.irGetObject
+import org.jetbrains.kotlin.ir.builders.irLong
 import org.jetbrains.kotlin.ir.builders.irReturn
 import org.jetbrains.kotlin.ir.builders.irSetField
 import org.jetbrains.kotlin.ir.builders.irString
@@ -69,7 +68,9 @@ import org.jetbrains.kotlin.ir.declarations.IrClass
 import org.jetbrains.kotlin.ir.declarations.IrDeclarationOrigin
 import org.jetbrains.kotlin.ir.declarations.IrEnumEntry
 import org.jetbrains.kotlin.ir.declarations.IrProperty
+import org.jetbrains.kotlin.ir.declarations.IrSimpleFunction
 import org.jetbrains.kotlin.ir.expressions.IrExpressionBody
+import org.jetbrains.kotlin.ir.expressions.impl.IrCallImpl
 import org.jetbrains.kotlin.ir.expressions.impl.IrConstImpl
 import org.jetbrains.kotlin.ir.expressions.impl.IrConstructorCallImpl
 import org.jetbrains.kotlin.ir.expressions.impl.IrExpressionBodyImpl
@@ -105,25 +106,18 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
             .symbol.createType(true, emptyList())
     private val realmObjectCompanionInterface =
         pluginContext.lookupClassOrThrow(REALM_MODEL_COMPANION)
-    private val table = pluginContext.lookupClassOrThrow(TABLE)
+    private val classInfoClass = pluginContext.lookupClassOrThrow(CLASS_INFO)
+    val classInfoCreateMethod = classInfoClass.lookupCompanionDeclaration<IrSimpleFunction>(CLASS_INFO_CREATE)
 
-    private val tableConstructor =
-        table.primaryConstructor ?: error("Couldn't find constructor for $TABLE")
-    private val classFlag: IrClass = pluginContext.lookupClassOrThrow(CLASS_FLAG)
-    private val classFlags =
-        classFlag.declarations.filterIsInstance<IrEnumEntry>()
-    private val propertyClass = pluginContext.lookupClassOrThrow(PROPERTY)
-    private val propertyConstructor =
-        propertyClass.primaryConstructor ?: error("Couldn't find constructor for $TABLE")
+    private val propertyClass = pluginContext.lookupClassOrThrow(PROPERTY_INFO)
+    val propertyCreateMethod = propertyClass.lookupCompanionDeclaration<IrSimpleFunction>(PROPERTY_INFO_CREATE)
+
     private val propertyType: IrClass = pluginContext.lookupClassOrThrow(PROPERTY_TYPE)
     private val propertyTypes =
         propertyType.declarations.filterIsInstance<IrEnumEntry>()
     private val collectionType: IrClass = pluginContext.lookupClassOrThrow(COLLECTION_TYPE)
     private val collectionTypes =
         collectionType.declarations.filterIsInstance<IrEnumEntry>()
-    private val propertyFlag: IrClass = pluginContext.lookupClassOrThrow(PROPERTY_FLAG)
-    private val propertyFlags =
-        propertyFlag.declarations.filterIsInstance<IrEnumEntry>()
 
     private val realmReferenceClass: IrClass = pluginContext.lookupClassOrThrow(REALM_REFERENCE)
     private val mediatorInterface: IrClass = pluginContext.lookupClassOrThrow(REALM_MEDIATOR_INTERFACE)
@@ -133,6 +127,10 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
         pluginContext.lookupClassOrThrow(FqNames.KOTLIN_COLLECTIONS_LIST)
     private val kProperty1Class: IrClass =
         pluginContext.lookupClassOrThrow(FqNames.KOTLIN_REFLECT_KPROPERTY1)
+    val realmClassImpl = pluginContext.lookupClassOrThrow(FqNames.REALM_CLASS_IMPL)
+    val realmClassCtor = pluginContext.lookupConstructorInClass(FqNames.REALM_CLASS_IMPL) {
+        it.owner.valueParameters.size == 2
+    }
 
     fun addProperties(irClass: IrClass): IrClass =
         irClass.apply {
@@ -140,7 +138,7 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
             addVariableProperty(realmModelInternalInterface, REALM_OWNER, realmReferenceClass.defaultType.makeNullable(), ::irNull)
             addVariableProperty(
                 realmModelInternalInterface,
-                OBJECT_TABLE_NAME,
+                OBJECT_CLASS_NAME,
                 pluginContext.irBuiltIns.stringType.makeNullable(),
                 ::irNull
             )
@@ -244,46 +242,44 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
         function.body = pluginContext.blockBody(function.symbol) {
             +irReturn(
                 IrConstructorCallImpl(
-                    startOffset,
-                    endOffset,
-                    type = table.defaultType,
-                    symbol = tableConstructor.symbol,
-                    constructorTypeArgumentsCount = 0,
+                    startOffset, endOffset,
+                    realmClassImpl.defaultType,
+                    realmClassCtor,
                     typeArgumentsCount = 0,
-                    valueArgumentsCount = 4
+                    constructorTypeArgumentsCount = 0,
+                    valueArgumentsCount = 2,
                 ).apply {
-                    var arg = 0
-                    // Name
-                    putValueArgument(arg++, irString(irClass.name.identifier))
-                    // Primary key
                     putValueArgument(
-                        arg++,
-                        if (primaryKey != null) irString(primaryKey) else {
-                            IrConstImpl.constNull(
-                                startOffset,
-                                endOffset,
-                                pluginContext.irBuiltIns.nothingNType
+                        0,
+                        IrCallImpl(
+                            startOffset,
+                            endOffset,
+                            type = classInfoClass.defaultType,
+                            symbol = classInfoCreateMethod.symbol,
+                            typeArgumentsCount = 0,
+                            valueArgumentsCount = 3
+                        ).apply {
+                            dispatchReceiver = irGetObject(classInfoClass.companionObject()!!.symbol)
+                            var arg = 0
+                            // Name
+                            putValueArgument(arg++, irString(irClass.name.identifier))
+                            // Primary key
+                            putValueArgument(
+                                arg++,
+                                if (primaryKey != null) irString(primaryKey) else {
+                                    IrConstImpl.constNull(
+                                        startOffset,
+                                        endOffset,
+                                        pluginContext.irBuiltIns.nothingNType
+                                    )
+                                }
                             )
+                            // num properties
+                            putValueArgument(arg++, irLong(fields.size.toLong()))
                         }
                     )
-                    // Flags
                     putValueArgument(
-                        arg++,
-                        buildSetOf(
-                            pluginContext, startOffset, endOffset, classFlag.defaultType,
-                            listOf(
-                                IrGetEnumValueImpl(
-                                    startOffset = UNDEFINED_OFFSET,
-                                    endOffset = UNDEFINED_OFFSET,
-                                    type = classFlag.defaultType,
-                                    symbol = classFlags.first { it.name == CLASS_FLAG_NORMAL }.symbol
-                                )
-                            )
-                        )
-                    )
-                    // Properties
-                    putValueArgument(
-                        arg++,
+                        1,
                         buildListOf(
                             pluginContext, startOffset, endOffset, propertyClass.defaultType,
                             fields.map { entry ->
@@ -328,16 +324,6 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
                                 val primaryKey = backingField.hasAnnotation(PRIMARY_KEY_ANNOTATION)
                                 val isIndexed = backingField.hasAnnotation(INDEX_ANNOTATION)
 
-                                val propertyFlags = mutableListOf<Name>()
-                                if (nullable) {
-                                    propertyFlags.add(PROPERTY_FLAG_NULLABLE)
-                                }
-                                if (primaryKey) {
-                                    propertyFlags.add(PROPERTY_FLAG_PRIMARY_KEY)
-                                }
-                                if (isIndexed) {
-                                    propertyFlags.add(PROPERTY_FLAG_INDEX)
-                                }
                                 val validPrimaryKeyTypes = with(pluginContext.irBuiltIns) {
                                     setOf(
                                         byteType,
@@ -362,15 +348,15 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
                                     )
                                 }
 
-                                IrConstructorCallImpl(
+                                IrCallImpl(
                                     startOffset,
                                     endOffset,
                                     type = propertyClass.defaultType,
-                                    symbol = propertyConstructor.symbol,
-                                    constructorTypeArgumentsCount = 0,
+                                    symbol = propertyCreateMethod.symbol,
                                     typeArgumentsCount = 0,
-                                    valueArgumentsCount = 7
+                                    valueArgumentsCount = 9
                                 ).apply {
+                                    dispatchReceiver = irGetObject(propertyClass.companionObject()!!.symbol)
                                     var arg = 0
                                     // Name
                                     putValueArgument(arg++, irString(entry.key))
@@ -425,17 +411,12 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
                                     )
                                     // Link property name
                                     putValueArgument(arg++, irString(""))
-                                    // Property flags
-                                    putValueArgument(
-                                        arg++,
-                                        buildSetOf(
-                                            pluginContext,
-                                            startOffset,
-                                            endOffset,
-                                            propertyFlag.defaultType,
-                                            propertyFlags(propertyFlags)
-                                        )
-                                    )
+                                    // isNullable
+                                    putValueArgument(arg++, irBoolean(nullable))
+                                    // isPrimaryKey
+                                    putValueArgument(arg++, irBoolean(primaryKey))
+                                    // isIndexed
+                                    putValueArgument(arg++, irBoolean(isIndexed))
                                 }
                             }
                         )
@@ -455,16 +436,6 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
 
     private fun getListType(generics: List<CoreType>?): PropertyType =
         checkNotNull(generics) { "Missing type for list." }[0].propertyType
-
-    private fun propertyFlags(flags: List<Name>): List<IrGetEnumValueImpl> =
-        flags.map { flag ->
-            IrGetEnumValueImpl(
-                startOffset = UNDEFINED_OFFSET,
-                endOffset = UNDEFINED_OFFSET,
-                type = propertyFlag.defaultType,
-                symbol = propertyFlags.first { flag == it.name }.symbol
-            )
-        }
 
     // Generate body for the synthetic new instance method defined inside the Companion instance previously declared via `RealmModelSyntheticCompanionExtension`
     fun addNewInstanceMethodBody(irClass: IrClass) {